refactor: features 구조 제시
