+++ src/features/assetLibrary/components/body/assetList/grid/GridItem.tsx
