+++ src/features/assetLibrary/components/body/assetList/toolbar/CategoryLabel.tsx
