+++ src/features/assetLibrary/components/body/category/Category.tsx
@@ -1,4 +1,4 @@
-import MAIN_CATEGORY_LIST from "@/features/constants/mainCategory";
+import MAIN_CATEGORY_LIST from "@/features/assetLibrary/constants/mainCategory";
 import styled from "styled-components";
 import CategoryItem from "./CategoryItem";
