+++ src/features/assetLibrary/index.tsx
@@ -1,0 +1,2 @@
+import AssetLibrary from "./components/AssetLibrary";
+export default AssetLibrary;
