+++ src/stories/assetLibrary/Body.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryFn, StoryObj } from "@storybook/react";
 import { basicColors } from "@/resources/colors/colors";
-import Body from "@/features/assetLibrary/body";
+import Body from "@/features/assetLibrary/components/body";
 
 const meta = {
   component: Body,
