+++ src/stories/assetLibrary/Footer.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import Footer from "@/features/assetLibrary/footer";
+import Footer from "@/features/assetLibrary/components/footer";
 
 const meta = {
   component: Footer,
