+++ src/stories/assetLibrary/Header.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryFn, StoryObj } from "@storybook/react";
-import Header from "@/features/assetLibrary/header";
+import Header from "@/features/assetLibrary/components/header";
 import { basicColors } from "@/resources/colors/colors";
 
 const meta = {
