+++ src/stories/assetLibrary/assetList/Toolbar/Toolbar.stories.tsx
@@ -1,4 +1,4 @@
-import Toolbar from "@/features/assetLibrary/body/assetList/toolbar";
+import Toolbar from "@/features/assetLibrary/components/body/assetList/toolbar";
 import type { Meta, StoryObj } from "@storybook/react";
 
 const meta = {
