+++ src/stories/assetLibrary/category/CategoryList.stories.tsx
@@ -1,5 +1,5 @@
-import CategoryItem from "@/features/assetLibrary/body/category/CategoryItem";
-import { MainCategory } from "@/features/constants/mainCategory";
+import CategoryItem from "@/features/assetLibrary/components/body/category/CategoryItem";
+import { MainCategory } from "@/features/assetLibrary/constants/mainCategory";
 import type { Meta, StoryObj } from "@storybook/react";
 
 const meta = {
